% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDNAx-pkg-deprecated.R
\name{gDNAx-pkg-deprecated}
\alias{gDNAx-pkg-deprecated}
\title{Deprecated functions in package 'gDNAx'}
\description{
The functions listed below are deprecated and provided for
compatibility with the previous version of `gDNAx` only, and will be defunct
at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
the replacement indicated below:
\itemize{
  \item{identifyStrandMode: \code{\link{strandedness,character-method}()}}
}
}
\keyword{internal}
