% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandedness.R
\name{identifyStrandMode}
\alias{identifyStrandMode}
\alias{strnessByFeature}
\title{Identify strandMode}
\usage{
identifyStrandMode(
  bfl,
  txdb,
  singleEnd = TRUE,
  stdChrom = TRUE,
  yieldSize = 1000000L,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)

strnessByFeature(
  bfl,
  features,
  singleEnd = TRUE,
  strandMode = 1L,
  yieldSize = 1000000L,
  ambiguous = FALSE,
  p = 0.6,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)
}
\arguments{
\item{bfl}{A \code{BamFile} or \code{BamFileList} object, or a character
string vector of BAM filenames.}

\item{txdb}{A character string of a \code{TxDb} package, or a \code{TxDb}
object, with gene and transcript annotations. For accurate calculations, it
is important that the version of these annotations matches the version of
the annotations used to inform the alignment of spliced reads, by the
short-read aligner software that generated the input BAM files.}

\item{singleEnd}{(Default FALSE) Logical value indicating if reads are
single (\code{TRUE}) or paired-end (\code{FALSE}).}

\item{stdChrom}{(Default TRUE) Logical value indicating whether only
alignments in the 'standard chromosomes' should be used. Consult the help
page of the function \code{\link[GenomeInfoDb]{keepStandardChromosomes}}
from the package \code{GenomeInfoDb} for further
information.}

\item{yieldSize}{(Default 5e5) Field inherited from
\code{\link[Rsamtools]{BamFile}}. The BAM is read by chunks.
\code{yieldSize} represents the number of records to read for each chunk.}

\item{verbose}{(Default TRUE) Logical value indicating if progress should be
reported through the execution of the code.}

\item{BPPARAM}{An object of a \linkS4class{BiocParallelParam} subclass
to configure the parallel execution of the code. By default, a
\linkS4class{SerialParam} object is used, which does not use any
parallelization, with the flag \code{progress=TRUE} to show progress
through the calculations.}

\item{features}{A \code{GRanges} or \code{GRangesList} object with
annotations of features (e.g. genes, transcripts, etc.).}

\item{strandMode}{(Default 1L) Numeric vector which can take values 0, 1,
or 2. The strand mode is a per-object switch on
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
objects that controls the behavior of the strand getter. See
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
class for further detail. If \code{singleEnd = TRUE}, then \code{strandMode}
is ignored.}

\item{ambiguous}{(Default FALSE) Logical value indicating if reads that
overlap a region with features annotated to both strands should be included
in the strandedness value computation.}

\item{p}{(Default 0.6) Numeric value for the exact binomial test performed
for the strandedness value of each feature, representing the hypothesized
probability of success (i.e. the strandedness value expected for a 
non-stranded dataset).}
}
\value{
A \link[base:list]{list} object with two elements:
\itemize{
  \item "strandMode": the \code{strandMode} of the sample(s) following
        \code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
        class definition. If all samples have the same \code{strandMode},
        the length of the vector is 1. It can take values: \code{NA}
        (library is not strand-specific), 1 (strand of pair is strand of 
        its first alignment), 2 (strand of pair is strand of its second 
        alignment) or "ambiguous" (additional category used here for 
        samples not fitting any of the three previous categories).
        See "Details" section below to know the classification criteria,
        as well as to how interpret results for single-end data.
  \item "Strandedness": data.frame with one row per sample and 3 columns.
        "strandMode1": proportion of alignments aligned to the same
        strand than a transcript according
        to the strand of its first alignment. "strandMode2": proportion
        of alignments aligned to the same strand than a transcript according
        to the strand of its second alignment. "ambiguous": alignments
        aligned to regions with transcripts in both strands.
}

A \linkS4class{SummarizedExperiment} with three assays:
\itemize{
  \item "strness": contains strandedness values for each feature and sample.
  \item "counts": number of reads aligning to each feature on the same 
         strand (according to \code{strandMode}).
  \item "counts_invstrand": number of reads aligning to each feature but on 
         the opposite strand (according to \code{strandMode}).
}
}
\description{
Identify \code{strandMode} (strandedness) in RNA-seq data samples based on
the proportion of reads aligning to the same or opposite strand as
transcripts in the annotations.

Compute strandedness for each feature in RNA-seq data samples based on
the proportion of reads aligning to the same strand as feature annotations
in relation to the total number of reads aligning to that feature.
}
\details{
If the value in the "strandMode1" column is > 0.90, \code{strandMode} is set
to 1L. If "strandMode2" column is > 0.90, \code{strandMode} is set to 2L. If
"strandMode1" and "strandMode2" are comprised between 0.40 and 0.60,
\code{strandMode} is set to \code{NA} If none of the three previous criteria
are met, \code{strandMode} is set to "ambiguous". This criteria can be
conservative in some cases (e.g. when there is genomic DNA contamination), 
for this reason we recommend to check the data.frame with strandedness 
values.
 
In case of single-end data, the same criteria are used, but the
interpretation of \code{strandMode = 1L} and \code{strandMode = 2L}
changes: when \code{strandMode = 1L} the strand of the read is concordant
with the reference annotations, when \code{strandMode = 2L} the correct 
read strand is the opposite to the one of the read.

A subset of 200,000 alignments overlapping gene annotations are used to
compute strandedness.

Strandedness is computed for each feature and BAM file according to the
\code{strandMode} specified in case of paired-end data. For single-end,
the original strand of reads is used. All alignments from the BAM file(s) 
are considered to compute the strandedness.

The \code{p} value should be close to 0.5, representing the strandedness
expected for a non-stranded RNA-seq library.
}
\examples{
library(gDNAinRNAseqData)

library(TxDb.Hsapiens.UCSC.hg38.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene

# Retrieving BAM files
bamfiles <- LiYu22subsetBAMfiles()
bamfiles <- bamfiles[c(1,7)] # using a subset of samples

strandM <- identifyStrandMode(bamfiles, txdb, singleEnd=FALSE)
strandM$strandMode
head(strandM$Strandedness)


features <- range(exonsBy(txdb, by="gene"))
features <- features[1:100]
sByFeature <- strnessByFeature(bamfiles, features, singleEnd=FALSE,
                            strandMode=2L)
sByFeature

}
