% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBAMtx.R
\name{filterBAMtx}
\alias{filterBAMtx}
\alias{filterBAMtxFlag}
\alias{testBAMtxFlag}
\title{Filter alignments in a BAM file using a transcriptome}
\usage{
filterBAMtx(
  object,
  path = ".",
  txflag = filterBAMtxFlag(),
  param = ScanBamParam(),
  yieldSize = 1e+06,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)

filterBAMtxFlag(
  isSpliceCompatibleJunction = FALSE,
  isSpliceCompatibleExonic = FALSE,
  isIntronic = FALSE,
  isIntergenic = FALSE
)

testBAMtxFlag(flag, value)
}
\arguments{
\item{object}{gDNAx object obtained with the function 'gDNAdx()'.}

\item{path}{Directory where to write the output BAM files.}

\item{txflag}{A value from a call to the function 'filterBAMtxFlag()'.}

\item{param}{A 'ScanBamParam' object.}

\item{yieldSize}{(Default 1e6) Number of records in the input BAM file to
yield each time the file is read. The lower the value, the smaller memory
consumption, but in the case of large BAM files, values below 1e6 records
may decrease the overall performance.}

\item{verbose}{(Default TRUE) Logical value indicating if progress should be
reported through the execution of the code.}

\item{BPPARAM}{An object of a \linkS4class{BiocParallelParam} subclass
to configure the parallel execution of the code. By default, a
\linkS4class{SerialParam} object is used, which does not use any
parallelization, with the flag \code{progress=TRUE} to show progress
through the calculations.}
}
\value{
A vector of output filename paths.
}
\description{
Filter alignments in a BAM file using criteria based on a
transcriptome annotation.

Use 'filterBAMtxFlag()' to set what types of alignment in a BAM 
file should be filtered using the function 'filterBAMtx()',
among being splice-compatible with one or more junctions,
splice-compatible exonic, intronic or intergenic.
}
