% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandedness.R
\name{strnessByFeature}
\alias{strnessByFeature}
\title{Compute strandedness for each feature}
\usage{
strnessByFeature(
  bfl,
  features,
  singleEnd = TRUE,
  strandMode = 1L,
  yieldSize = 1000000L,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)
}
\arguments{
\item{bfl}{A \code{BamFile} or \code{BamFileList} object, or a character
string vector of BAM filenames.}

\item{features}{A \code{GRanges} or \code{GRangesList} object with
annotations of features (e.g. genes, transcripts, etc.).}

\item{singleEnd}{(Default FALSE) Logical value indicating if reads are
single (\code{TRUE}) or paired-end (\code{FALSE}).}

\item{strandMode}{(Default 1L) Numeric vector which can take values 0, 1,
or 2. The strand mode is a per-object switch on
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
objects that controls the behavior of the strand getter. See
\code{\link[GenomicAlignments:GAlignmentPairs-class]{GAlignmentPairs}}
class for further detail. If \code{singleEnd = TRUE}, then \code{strandMode}
is ignored.}

\item{yieldSize}{(Default 5e5) Field inherited from
\code{\link[Rsamtools]{BamFile}}. The BAM is read by chunks.
\code{yieldSize} represents the number of records to read for each chunk.}

\item{verbose}{(Default TRUE) Logical value indicating if progress should be
reported through the execution of the code.}

\item{BPPARAM}{An object of a \linkS4class{BiocParallelParam} subclass
to configure the parallel execution of the code. By default, a
\linkS4class{SerialParam} object is used, which does not use any
parallelization, with the flag \code{progress=TRUE} to show progress
through the calculations.}
}
\value{
A \linkS4class{SummarizedExperiment} with three assays:
\itemize{
  \item "strness": contains strandedness values for each feature and sample.
  \item "counts": number of reads aligning to each feature on the same 
         strand (according to \code{strandMode}).
  \item "counts_invstrand": number of reads aligning to each feature but on 
         the opposite strand (according to \code{strandMode}).
}
}
\description{
Compute strandedness for each feature in RNA-seq data samples based on
the proportion of reads aligning to the same strand as feature annotations
in relation to the total number of reads aligning to that feature.
}
\details{
Strandedness is computed for each feature and BAM file according to the
\code{strandMode} specified in case of paired-end data. For single-end,
the original strand of reads is used. All alignments from the BAM file(s) 
are considered to compute the strandedness.
}
